Struct client_sdk::api::RateLimit

pub struct RateLimit {
    pub emission_interval: Duration,
    pub burst_size: u64,
}

Rate-limiting configuration for a command

Fields
emission_interval: Duration

Ideal time between each request

burst_size: u64

Maximum number of requests that can be made in a burst, before rate-limiting kicks in.

Implementations

impl RateLimit

pub const DEFAULT: RateLimit = _

Default rate-limit config for commands when not otherwise specified.

RateLimit {
    emission_interval: 50ms,
    burst_size: 5,
}

Or approximately 20 requests per second, with up to 5 requests burst within the emission_interval, but the client must wait for them to replenish before another burst.