Trait client_sdk::api::Command

pub trait Command: Sealed {
    type Result: DeserializeOwned;
    type Body: Serialize;

    const METHOD: Method;
    const FLAGS: CommandFlags;

    fn format_path<W: Write>(&self, w: W) -> Result;

    fn body(&self) -> &Self::Body;

    fn perms(&self) -> Permission;

    fn schema(gen: &mut SchemaGenerator) -> (String, PathItem);

    fn body_size_hint(&self) -> usize { ... }

    fn add_headers(&self, _map: &mut HeaderMap) { ... }
}

Client Command, tells the client to perform specific requests

A “Command” is a mid-level abstraction around REST endpoints and their bodies. Not perfect, but zero-cost and simple. Other abstractions can be built on top of it.

A command consists of three parts: the URL, the “body”, and headers.

For the case of GET/OPTIONS commands, the body becomes query parameters.

Associated Types

type Result: DeserializeOwned

Object returned from the server as the result of a command

type Body: Serialize

Associated Constants

const METHOD: Method

HTTP Method used to execute the command

const FLAGS: CommandFlags

Required methods

fn format_path<W: Write>(&self, w: W) -> Result

Serialize/format the REST path (without query)

fn body(&self) -> &Self::Body

fn perms(&self) -> Permission

Computes required permissions

fn schema(gen: &mut SchemaGenerator) -> (String, PathItem)

Provided methods

fn body_size_hint(&self) -> usize

Hint given to preallocate body size, only used for query strings

fn add_headers(&self, _map: &mut HeaderMap)

Insert any additional headers required to perform this command