Module client_sdk::models

Object data structures used within Lantern

Re-exports

pub use nullable::Nullable;

pub use self::asset::*;

pub use self::auth::*;

pub use self::config::*;

pub use self::embed::*;

pub use self::emote::*;

pub use self::file::*;

pub use self::gateway::*;

pub use self::invite::*;

pub use self::message::*;

pub use self::party::*;

pub use self::permission::*;

pub use self::presence::*;

pub use self::role::*;

pub use self::room::*;

pub use self::session::*;

pub use self::sf::*;

pub use self::stats::*;

pub use self::thread::*;

pub use self::user::*;

Modules

asset

auth

config

embed

emote

file

gateway

invite

message

nullable

party

permission

presence

role

room

session

sf

stats

thread

user

Structs

Arc

An atomically reference counted shared pointer

SmolStr

A SmolStr is a string type that has the following properties:

ThinVec

See the crate’s top level documentation for a description of this type.

Timestamp

UTC Timestamp with nanosecond precision, millisecond-precision when serialized to serde (JSON).

Enums

ArchivedCursor

An archived Cursor

Cursor

Directional search query

CursorResolver

The resolver for an archived Cursor