Enum client_sdk::models::gateway::message::ClientMsg

pub enum ClientMsg {
    Heartbeat {
        op: ClientMsgOpcode,
        payload: HeartbeatPayload,
    },
    Identify {
        op: ClientMsgOpcode,
        payload: IdentifyPayload,
    },
    Resume {
        op: ClientMsgOpcode,
        payload: ResumePayload,
    },
    SetPresence {
        op: ClientMsgOpcode,
        payload: SetPresencePayload,
    },
    Subscribe {
        op: ClientMsgOpcode,
        payload: SubscribePayload,
    },
    Unsubscribe {
        op: ClientMsgOpcode,
        payload: UnsubscribePayload,
    },
}

Messages sent from the client to the server

Variants

Heartbeat

Fields

op: ClientMsgOpcode

payload: HeartbeatPayload

See new_heartbeat for an easy way to create this message.

Identify

Fields

op: ClientMsgOpcode

payload: IdentifyPayload

See new_identify for an easy way to create this message.

Resume

Fields

op: ClientMsgOpcode

payload: ResumePayload

See new_resume for an easy way to create this message.

SetPresence

Fields

op: ClientMsgOpcode

payload: SetPresencePayload

See new_set_presence for an easy way to create this message.

Subscribe

Fields

op: ClientMsgOpcode

payload: SubscribePayload

See new_subscribe for an easy way to create this message.

Unsubscribe

Fields

op: ClientMsgOpcode

payload: UnsubscribePayload

See new_unsubscribe for an easy way to create this message.

Implementations

impl ClientMsg

pub const fn heartbeat(payload: HeartbeatPayload) -> Self

Create new Heartbeat message from raw payload struct.

pub fn new_heartbeat() -> Self

Create new Heartbeat message from payload fields.

pub const fn identify(payload: IdentifyPayload) -> Self

Create new Identify message from raw payload struct.

pub fn new_identify(inner: impl Into<Box<Identify>>) -> Self

Create new Identify message from payload fields.

pub const fn resume(payload: ResumePayload) -> Self

Create new Resume message from raw payload struct.

pub fn new_resume(session: impl Into<Snowflake>) -> Self

Create new Resume message from payload fields.

pub const fn set_presence(payload: SetPresencePayload) -> Self

Create new SetPresence message from raw payload struct.

pub fn new_set_presence(inner: impl Into<Box<SetPresence>>) -> Self

Create new SetPresence message from payload fields.

pub const fn subscribe(payload: SubscribePayload) -> Self

Create new Subscribe message from raw payload struct.

pub fn new_subscribe(party_id: impl Into<Snowflake>) -> Self

Create new Subscribe message from payload fields.

pub const fn unsubscribe(payload: UnsubscribePayload) -> Self

Create new Unsubscribe message from raw payload struct.

pub fn new_unsubscribe(party_id: impl Into<Snowflake>) -> Self

Create new Unsubscribe message from payload fields.