Enum client_sdk::models::gateway::message::ServerMsg

pub enum ServerMsg {

    Hello {
        op: ServerMsgOpcode,
        payload: HelloPayload,
    },
    HeartbeatAck {
        op: ServerMsgOpcode,
        payload: HeartbeatAckPayload,
    },
    Ready {
        op: ServerMsgOpcode,
        payload: ReadyPayload,
    },
    InvalidSession {
        op: ServerMsgOpcode,
        payload: InvalidSessionPayload,
    },
    PartyCreate {
        op: ServerMsgOpcode,
        payload: PartyCreatePayload,
    },
    PartyUpdate {
        op: ServerMsgOpcode,
        payload: PartyUpdatePayload,
    },
    PartyDelete {
        op: ServerMsgOpcode,
        payload: PartyDeletePayload,
    },
    RoleCreate {
        op: ServerMsgOpcode,
        payload: RoleCreatePayload,
    },
    RoleUpdate {
        op: ServerMsgOpcode,
        payload: RoleUpdatePayload,
    },
    RoleDelete {
        op: ServerMsgOpcode,
        payload: RoleDeletePayload,
    },
    MemberAdd {
        op: ServerMsgOpcode,
        payload: MemberAddPayload,
    },
    MemberUpdate {
        op: ServerMsgOpcode,
        payload: MemberUpdatePayload,
    },
    MemberRemove {
        op: ServerMsgOpcode,
        payload: MemberRemovePayload,
    },
    MemberBan {
        op: ServerMsgOpcode,
        payload: MemberBanPayload,
    },
    MemberUnban {
        op: ServerMsgOpcode,
        payload: MemberUnbanPayload,
    },
    RoomCreate {
        op: ServerMsgOpcode,
        payload: RoomCreatePayload,
    },
    RoomUpdate {
        op: ServerMsgOpcode,
        payload: RoomUpdatePayload,
    },
    RoomDelete {
        op: ServerMsgOpcode,
        payload: RoomDeletePayload,
    },
    RoomPinsUpdate {
        op: ServerMsgOpcode,
        payload: RoomPinsUpdatePayload,
    },
    MessageCreate {
        op: ServerMsgOpcode,
        payload: MessageCreatePayload,
    },
    MessageUpdate {
        op: ServerMsgOpcode,
        payload: MessageUpdatePayload,
    },
    MessageDelete {
        op: ServerMsgOpcode,
        payload: MessageDeletePayload,
    },
    MessageReactionAdd {
        op: ServerMsgOpcode,
        payload: MessageReactionAddPayload,
    },
    MessageReactionRemove {
        op: ServerMsgOpcode,
        payload: MessageReactionRemovePayload,
    },
    MessageReactionRemoveAll {
        op: ServerMsgOpcode,
        payload: MessageReactionRemoveAllPayload,
    },
    MessageReactionRemoveEmote {
        op: ServerMsgOpcode,
        payload: MessageReactionRemoveEmotePayload,
    },
    PresenceUpdate {
        op: ServerMsgOpcode,
        payload: PresenceUpdatePayload,
    },
    TypingStart {
        op: ServerMsgOpcode,
        payload: TypingStartPayload,
    },
    UserUpdate {
        op: ServerMsgOpcode,
        payload: UserUpdatePayload,
    },

}

Messages send from the server to the client

Variants

Hello

Fields

op: ServerMsgOpcode

payload: HelloPayload

The Hello message initialates the gateway session and expects a ClientMsg::Identify in return.

See new_hello for an easy way to create this message.

HeartbeatAck

Fields

op: ServerMsgOpcode

payload: HeartbeatAckPayload

Acknowledgement of a heartbeat

See new_heartbeat_ack for an easy way to create this message.

Ready

Fields

op: ServerMsgOpcode

payload: ReadyPayload

See new_ready for an easy way to create this message.

InvalidSession

Fields

op: ServerMsgOpcode

payload: InvalidSessionPayload

Sent when the session is no longer valid

See new_invalid_session for an easy way to create this message.

PartyCreate

Fields

op: ServerMsgOpcode

payload: PartyCreatePayload

See new_party_create for an easy way to create this message.

PartyUpdate

Fields

op: ServerMsgOpcode

payload: PartyUpdatePayload

See new_party_update for an easy way to create this message.

PartyDelete

Fields

op: ServerMsgOpcode

payload: PartyDeletePayload

See new_party_delete for an easy way to create this message.

RoleCreate

Fields

op: ServerMsgOpcode

payload: RoleCreatePayload

See new_role_create for an easy way to create this message.

RoleUpdate

Fields

op: ServerMsgOpcode

payload: RoleUpdatePayload

See new_role_update for an easy way to create this message.

RoleDelete

Fields

op: ServerMsgOpcode

payload: RoleDeletePayload

See new_role_delete for an easy way to create this message.

MemberAdd

Fields

op: ServerMsgOpcode

payload: MemberAddPayload

See new_member_add for an easy way to create this message.

MemberUpdate

Fields

op: ServerMsgOpcode

payload: MemberUpdatePayload

See new_member_update for an easy way to create this message.

MemberRemove

Fields

op: ServerMsgOpcode

payload: MemberRemovePayload

See new_member_remove for an easy way to create this message.

MemberBan

Fields

op: ServerMsgOpcode

payload: MemberBanPayload

See new_member_ban for an easy way to create this message.

MemberUnban

Fields

op: ServerMsgOpcode

payload: MemberUnbanPayload

See new_member_unban for an easy way to create this message.

RoomCreate

Fields

op: ServerMsgOpcode

payload: RoomCreatePayload

See new_room_create for an easy way to create this message.

RoomUpdate

Fields

op: ServerMsgOpcode

payload: RoomUpdatePayload

See new_room_update for an easy way to create this message.

RoomDelete

Fields

op: ServerMsgOpcode

payload: RoomDeletePayload

See new_room_delete for an easy way to create this message.

RoomPinsUpdate

Fields

op: ServerMsgOpcode

payload: RoomPinsUpdatePayload

See new_room_pins_update for an easy way to create this message.

MessageCreate

Fields

op: ServerMsgOpcode

payload: MessageCreatePayload

See new_message_create for an easy way to create this message.

MessageUpdate

Fields

op: ServerMsgOpcode

payload: MessageUpdatePayload

See new_message_update for an easy way to create this message.

MessageDelete

Fields

op: ServerMsgOpcode

payload: MessageDeletePayload

See new_message_delete for an easy way to create this message.

MessageReactionAdd

Fields

op: ServerMsgOpcode

payload: MessageReactionAddPayload

See new_message_reaction_add for an easy way to create this message.

MessageReactionRemove

Fields

op: ServerMsgOpcode

payload: MessageReactionRemovePayload

See new_message_reaction_remove for an easy way to create this message.

MessageReactionRemoveAll

Fields

op: ServerMsgOpcode

payload: MessageReactionRemoveAllPayload

See new_message_reaction_remove_all for an easy way to create this message.

MessageReactionRemoveEmote

Fields

op: ServerMsgOpcode

payload: MessageReactionRemoveEmotePayload

See new_message_reaction_remove_emote for an easy way to create this message.

PresenceUpdate

Fields

op: ServerMsgOpcode

payload: PresenceUpdatePayload

See new_presence_update for an easy way to create this message.

TypingStart

Fields

op: ServerMsgOpcode

payload: TypingStartPayload

See new_typing_start for an easy way to create this message.

UserUpdate

Fields

op: ServerMsgOpcode

payload: UserUpdatePayload

See new_user_update for an easy way to create this message.

Implementations

impl ServerMsg

pub const fn hello(payload: HelloPayload) -> Self

Create new Hello message from raw payload struct.

The Hello message initialates the gateway session and expects a ClientMsg::Identify in return.

pub fn new_hello(inner: impl Into<Hello>) -> Self

Create new Hello message from payload fields.

The Hello message initialates the gateway session and expects a ClientMsg::Identify in return.

pub const fn heartbeat_ack(payload: HeartbeatAckPayload) -> Self

Create new HeartbeatAck message from raw payload struct.

Acknowledgement of a heartbeat

pub fn new_heartbeat_ack() -> Self

Create new HeartbeatAck message from payload fields.

Acknowledgement of a heartbeat

pub const fn ready(payload: ReadyPayload) -> Self

Create new Ready message from raw payload struct.

pub fn new_ready(inner: impl Into<Box<Ready>>) -> Self

Create new Ready message from payload fields.

pub const fn invalid_session(payload: InvalidSessionPayload) -> Self

Create new InvalidSession message from raw payload struct.

Sent when the session is no longer valid

pub fn new_invalid_session() -> Self

Create new InvalidSession message from payload fields.

Sent when the session is no longer valid

pub const fn party_create(payload: PartyCreatePayload) -> Self

Create new PartyCreate message from raw payload struct.

pub fn new_party_create(inner: impl Into<Box<Party>>) -> Self

Create new PartyCreate message from payload fields.

pub const fn party_update(payload: PartyUpdatePayload) -> Self

Create new PartyUpdate message from raw payload struct.

pub fn new_party_update(inner: impl Into<Box<PartyUpdateEvent>>) -> Self

Create new PartyUpdate message from payload fields.

pub const fn party_delete(payload: PartyDeletePayload) -> Self

Create new PartyDelete message from raw payload struct.

pub fn new_party_delete(id: impl Into<Snowflake>) -> Self

Create new PartyDelete message from payload fields.

pub const fn role_create(payload: RoleCreatePayload) -> Self

Create new RoleCreate message from raw payload struct.

pub fn new_role_create(inner: impl Into<Box<Role>>) -> Self

Create new RoleCreate message from payload fields.

pub const fn role_update(payload: RoleUpdatePayload) -> Self

Create new RoleUpdate message from raw payload struct.

pub fn new_role_update(inner: impl Into<Box<Role>>) -> Self

Create new RoleUpdate message from payload fields.

pub const fn role_delete(payload: RoleDeletePayload) -> Self

Create new RoleDelete message from raw payload struct.

pub fn new_role_delete(inner: impl Into<Box<RoleDeleteEvent>>) -> Self

Create new RoleDelete message from payload fields.

pub const fn member_add(payload: MemberAddPayload) -> Self

Create new MemberAdd message from raw payload struct.

pub fn new_member_add(inner: impl Into<Box<PartyMemberEvent>>) -> Self

Create new MemberAdd message from payload fields.

pub const fn member_update(payload: MemberUpdatePayload) -> Self

Create new MemberUpdate message from raw payload struct.

pub fn new_member_update(inner: impl Into<Box<PartyMemberEvent>>) -> Self

Create new MemberUpdate message from payload fields.

pub const fn member_remove(payload: MemberRemovePayload) -> Self

Create new MemberRemove message from raw payload struct.

pub fn new_member_remove(inner: impl Into<Arc<PartyMemberEvent>>) -> Self

Create new MemberRemove message from payload fields.

pub const fn member_ban(payload: MemberBanPayload) -> Self

Create new MemberBan message from raw payload struct.

pub fn new_member_ban(inner: impl Into<Arc<PartyMemberEvent>>) -> Self

Create new MemberBan message from payload fields.

pub const fn member_unban(payload: MemberUnbanPayload) -> Self

Create new MemberUnban message from raw payload struct.

pub fn new_member_unban(inner: impl Into<Box<PartyMemberEvent>>) -> Self

Create new MemberUnban message from payload fields.

pub const fn room_create(payload: RoomCreatePayload) -> Self

Create new RoomCreate message from raw payload struct.

pub fn new_room_create(room: impl Into<Box<()>>) -> Self

Create new RoomCreate message from payload fields.

pub const fn room_update(payload: RoomUpdatePayload) -> Self

Create new RoomUpdate message from raw payload struct.

pub fn new_room_update(room: impl Into<Box<()>>) -> Self

Create new RoomUpdate message from payload fields.

pub const fn room_delete(payload: RoomDeletePayload) -> Self

Create new RoomDelete message from raw payload struct.

pub fn new_room_delete(room: impl Into<Box<RoomDeleteEvent>>) -> Self

Create new RoomDelete message from payload fields.

pub const fn room_pins_update(payload: RoomPinsUpdatePayload) -> Self

Create new RoomPinsUpdate message from raw payload struct.

pub fn new_room_pins_update() -> Self

Create new RoomPinsUpdate message from payload fields.

pub const fn message_create(payload: MessageCreatePayload) -> Self

Create new MessageCreate message from raw payload struct.

pub fn new_message_create(msg: impl Into<Box<RoomMessage>>) -> Self

Create new MessageCreate message from payload fields.

pub const fn message_update(payload: MessageUpdatePayload) -> Self

Create new MessageUpdate message from raw payload struct.

pub fn new_message_update(msg: impl Into<Box<RoomMessage>>) -> Self

Create new MessageUpdate message from payload fields.

pub const fn message_delete(payload: MessageDeletePayload) -> Self

Create new MessageDelete message from raw payload struct.

pub fn new_message_delete(msg: impl Into<Box<MessageDeleteEvent>>) -> Self

Create new MessageDelete message from payload fields.

pub const fn message_reaction_add(payload: MessageReactionAddPayload) -> Self

Create new MessageReactionAdd message from raw payload struct.

pub fn new_message_reaction_add() -> Self

Create new MessageReactionAdd message from payload fields.

pub const fn message_reaction_remove(
payload: MessageReactionRemovePayload
) -> Self

Create new MessageReactionRemove message from raw payload struct.

pub fn new_message_reaction_remove() -> Self

Create new MessageReactionRemove message from payload fields.

pub const fn message_reaction_remove_all(
payload: MessageReactionRemoveAllPayload
) -> Self

Create new MessageReactionRemoveAll message from raw payload struct.

pub fn new_message_reaction_remove_all() -> Self

Create new MessageReactionRemoveAll message from payload fields.

pub const fn message_reaction_remove_emote(
payload: MessageReactionRemoveEmotePayload
) -> Self

Create new MessageReactionRemoveEmote message from raw payload struct.

pub fn new_message_reaction_remove_emote() -> Self

Create new MessageReactionRemoveEmote message from payload fields.

pub const fn presence_update(payload: PresenceUpdatePayload) -> Self

Create new PresenceUpdate message from raw payload struct.

pub fn new_presence_update(
party: impl Into<Option<Snowflake>>,
inner: impl Into<Arc<UserPresenceEvent>>
) -> Self

Create new PresenceUpdate message from payload fields.

pub const fn typing_start(payload: TypingStartPayload) -> Self

Create new TypingStart message from raw payload struct.

pub fn new_typing_start(t: impl Into<Box<TypingStart>>) -> Self

Create new TypingStart message from payload fields.

pub const fn user_update(payload: UserUpdatePayload) -> Self

Create new UserUpdate message from raw payload struct.

pub fn new_user_update(user: impl Into<Arc<User>>) -> Self

Create new UserUpdate message from payload fields.

impl ServerMsg

pub fn matching_intent(&self) -> Option<Intent>