client_sdk::models::gateway
Struct Intent

pub struct Intent(/* private fields */);

Implementations

impl Intent

pub const PARTIES: Self = _

PARTY_CREATE

PARTY_UPDATE

PARTY_DELETE

PARTY_ROLE_CREATE

PARTY_ROLE_UPDATE

PARTY_ROLE_DELETE

CHANNEL_CREATE

CHANNEL_UPDATE

CHANNEL_DELETE

CHANNEL_PINS_UPDATE

pub const PARTY_MEMBERS: Self = _

PARTY_MEMBER_ADD

PARTY_MEMBER_UPDATE

PARTY_MEMBER_REMOVE

pub const PARTY_BANS: Self = _

PARTY_BAN_ADD

PARTY_BAN_REMOVE

pub const PARTY_EMOTES: Self = _

PARTY_EMOJIS_UPDATE

pub const PARTY_INTEGRATIONS: Self = _

PARTY_INTEGRATIONS_UPDATE

INTEGRATION_CREATE

INTEGRATION_UPDATE

INTEGRATION_DELETE

pub const PARTY_WEBHOOKS: Self = _

WEBHOOKS_UPDATE

pub const PARTY_INVITES: Self = _

INVITE_CREATE

INVITE_DELETE

pub const VOICE_STATUS: Self = _

VOICE_STATE_UPDATE

pub const PRESENCE: Self = _

PRESENCE_UPDATE

pub const MESSAGES: Self = _

MESSAGE_CREATE

MESSAGE_UPDATE

MESSAGE_DELETE

MESSAGE_DELETE_BULK

pub const MESSAGE_REACTIONS: Self = _

MESSAGE_REACTION_ADD

MESSAGE_REACTION_REMOVE

MESSAGE_REACTION_REMOVE_ALL

MESSAGE_REACTION_REMOVE_EMOTE

pub const MESSAGE_TYPING: Self = _

TYPING_START

pub const DIRECT_MESSAGES: Self = _

MESSAGE_CREATE

MESSAGE_UPDATE

MESSAGE_DELETE

CHANNEL_PINS_UPDATE

pub const DIRECT_MESSAGE_REACTIONS: Self = _

MESSAGE_REACTION_ADD

MESSAGE_REACTION_REMOVE

MESSAGE_REACTION_REMOVE_ALL

MESSAGE_REACTION_REMOVE_EMOTE

pub const DIRECT_MESSAGE_TYPING: Self = _

TYPING_START

pub const PROFILE_UPDATES: Self = _

impl Intent

pub const fn empty() -> Self

Get a flags value with all bits unset.

pub const fn all() -> Self

Get a flags value with all known bits set.

pub const fn bits(&self) -> u32

Get the underlying bits value.

The returned value is exactly the bits set in this flags value.

pub const fn from_bits(bits: u32) -> Option<Self>

Convert from a bits value.

This method will return None if any unknown bits are set.

pub const fn from_bits_truncate(bits: u32) -> Self

Convert from a bits value, unsetting any unknown bits.

pub const fn from_bits_retain(bits: u32) -> Self

Convert from a bits value exactly.

pub fn from_name(name: &str) -> Option<Self>

Get a flags value with the bits of a flag with the given name set.

This method will return None if name is empty or doesn’t correspond to any named flag.

pub const fn is_empty(&self) -> bool

Whether all bits in this flags value are unset.

pub const fn is_all(&self) -> bool

Whether all known bits in this flags value are set.

pub const fn intersects(&self, other: Self) -> bool

Whether any set bits in a source flags value are also set in a target flags value.

pub const fn contains(&self, other: Self) -> bool

Whether all set bits in a source flags value are also set in a target flags value.

pub fn insert(&mut self, other: Self)

The bitwise or (|) of the bits in two flags values.

pub fn remove(&mut self, other: Self)

The intersection of a source flags value with the complement of a target flags value (&!).

This method is not equivalent to self & !other when other has unknown bits set. remove won’t truncate other, but the ! operator will.

pub fn toggle(&mut self, other: Self)

The bitwise exclusive-or (^) of the bits in two flags values.

pub fn set(&mut self, other: Self, value: bool)

Call insert when value is true or remove when value is false.

pub const fn intersection(self, other: Self) -> Self

The bitwise and (&) of the bits in two flags values.

pub const fn union(self, other: Self) -> Self

The bitwise or (|) of the bits in two flags values.

pub const fn difference(self, other: Self) -> Self

The intersection of a source flags value with the complement of a target flags value (&!).

This method is not equivalent to self & !other when other has unknown bits set. difference won’t truncate other, but the ! operator will.

pub const fn symmetric_difference(self, other: Self) -> Self

The bitwise exclusive-or (^) of the bits in two flags values.

pub const fn complement(self) -> Self

The bitwise negation (!) of the bits in a flags value, truncating the result.

impl Intent

pub const fn iter(&self) -> Iter<Intent>

Yield a set of contained flags values.

Each yielded flags value will correspond to a defined named flag. Any unknown bits will be yielded together as a final flags value.

pub const fn iter_names(&self) -> IterNames<Intent>

Yield a set of contained named flags values.

This method is like iter, except only yields bits in contained named flags. Any unknown bits, or bits not corresponding to a contained flag will not be yielded.