Module client_sdk::models::message

Structs

ArchivedAttachment

An archived Attachment

ArchivedMessage

An archived Message

ArchivedMessageFlags

Archived version of MessageFlags, automatically made endian-agnostic. When comparing with the native bitflags, the archived bitflags will truncate any extra bits.

ArchivedReactionFull

An archived ReactionFull

ArchivedReactionShorthand

An archived ReactionShorthand

Attachment

AttachmentResolver

The resolver for an archived Attachment

Message

MessageFlags

MessageResolver

The resolver for an archived Message

ReactionFull

ReactionFullResolver

The resolver for an archived ReactionFull

ReactionShorthand

ReactionShorthandResolver

The resolver for an archived ReactionShorthand

Enums

ArchivedEmoteOrEmoji

An archived EmoteOrEmoji

ArchivedMessageKind

ArchivedReaction

An archived Reaction

EmoteOrEmoji

Simple enum for custom emote ids or emoji symbols

EmoteOrEmojiResolver

The resolver for an archived EmoteOrEmoji

MessageKind

Reaction

ReactionResolver

The resolver for an archived Reaction