client_sdk::models::nullable
Enum Nullable

#[repr(u8)]
pub enum Nullable<T> {
    Undefined = 0,
    Null = 1,
    Some(T),
}

Similar to Option, Undefined can be used when data may exist but is not provided.

Consider using Nullable when you want to distinguish between None and Some(None), as there is a difference between not providing a value and providing a null value.

This is commonly used when setting a field or property to null/None versus not changing it all, such as in patch requests where you want to update only certain fields. Additionally, when data is not available to the user due to permissions, it may also be set to Undefined. It may still exist, but its state is not known.

This is not equivalent to three-state boolean logic, and Undefined values will be considered equal to each other.

Variants

Undefined = 0

Neither present nor absent, an indeterminant value.

Null = 1

Certainly absent of value.

Some(T)

Certainly present of value.

Implementations

impl<T> Nullable<T>

pub const fn is_undefined(&self) -> bool

Returns true if the value is Undefined.

pub const fn is_null(&self) -> bool

Returns true if the value is Null.

pub const fn is_some(&self) -> bool

Returns true if the value is Some.

pub fn map<F, U>(self, f: F) -> Nullable<U>
where F: FnOnce(T) -> U,

Maps an inner Some value to a different value, carrying over Null and Undefined unchanged.

pub fn as_ref(&self) -> Nullable<&T>

Converts Nullable<T> to Nullable<&T>.

pub fn as_mut(&mut self) -> Nullable<&mut T>

Converts Nullable<T> to Nullable<&mut T>.

pub fn convert<U>(self) -> Nullable<U>
where T: Into<U>,

Maps an inner Some value to a different value, using Into.

Equivalent to .map(Into::into).