client_sdk::models
Struct Timestamp

pub struct Timestamp(/* private fields */);

UTC Timestamp with nanosecond precision, millisecond-precision when serialized to serde (JSON).

A Deref/DerefMut implementation is provided to gain access to the inner PrimitiveDateTime object.

Implementations

impl Timestamp

pub fn to_calendar_date(&self) -> (i32, Month, u8)

Get the year, month, and day.

Like PrimitiveDateTime::to_calendar_date, but optimized for SSE2/AVX2 when available.

assert_eq!(
    Timestamp::from(datetime!(2019-01-01 0:00)).to_calendar_date(),
    (2019, Month::January, 1)
);

impl Timestamp

pub fn now_utc() -> Timestamp

Get the current time, assuming UTC.

impl Timestamp

pub const UNIX_EPOCH: Timestamp = _

Unix Epoch – 1970-01-01 Midnight

pub const fn from_primitive_datetime(dt: PrimitiveDateTime) -> Timestamp

Constructs a Timestamp from a [PrimitiveDateTime]

pub fn duration_since(self, earlier: Timestamp) -> Duration

Returns the amount of time elapsed from an earlier point in time.

pub fn format_raw<F, O, P>( &self, offset: UtcOffset, ) -> TimestampStr<FormatString<F, O, P>>
where F: Bit, O: Bit, P: Unsigned, FormatString<F, O, P>: IsValidFormat,

Formats the timestamp given the provided formatting parameters

pub fn format_with_precision<P>(&self) -> TimestampStr<FormatString<B1, B0, P>>
where P: Unsigned, FormatString<B1, B0, P>: IsValidFormat,

Formats a full timestamp without offset, using the given subsecond precision level.

pub fn format( &self, ) -> TimestampStr<FormatString<B1, B0, UInt<UInt<UTerm, B1>, B1>>>

Format timestamp to ISO8601 with full punctuation, to millisecond precision.

pub fn format_nanoseconds( &self, ) -> TimestampStr<FormatString<B1, B0, UInt<UInt<UInt<UInt<UTerm, B1>, B0>, B0>, B1>>>

Format timestamp to ISO8601 with extended precision to nanoseconds.

pub fn format_microseconds( &self, ) -> TimestampStr<FormatString<B1, B0, UInt<UInt<UInt<UTerm, B1>, B1>, B0>>>

Format timestamp to ISO8601 with extended precision to microseconds.

pub fn format_short( &self, ) -> TimestampStr<FormatString<B0, B0, UInt<UInt<UTerm, B1>, B1>>>

Format timestamp to ISO8601 without most punctuation, to millisecond precision.

pub fn format_with_offset( &self, offset: UtcOffset, ) -> TimestampStr<FormatString<B1, B1, UInt<UInt<UTerm, B1>, B1>>>

Format timestamp to ISO8601 with arbitrary UTC offset. Any offset is formatted as +HH:MM, and no timezone conversions are done. It is interpreted literally.

pub fn format_with_offset_and_precision<P>( &self, offset: UtcOffset, ) -> TimestampStr<FormatString<B1, B1, P>>
where P: Unsigned, FormatString<B1, B1, P>: IsValidFormat,

Formats a full timestamp with timezone offset, and the provided level of subsecond precision.

pub fn parse(ts: &str) -> Option<Timestamp>

Parse to UTC timestamp from any ISO8601 string. Offsets are applied during parsing.

pub const fn assume_offset(self, offset: UtcOffset) -> OffsetDateTime

Convert to time::OffsetDateTime with the given offset.

pub const fn checked_add(self, duration: Duration) -> Option<Timestamp>

Computes self + duration, returning None if an overflow occurred.

See [PrimitiveDateTime::checked_add] for more implementation details

pub const fn checked_sub(self, duration: Duration) -> Option<Timestamp>

Computes self - duration, returning None if an overflow occurred.

See [PrimitiveDateTime::checked_sub] for more implementation details

pub const fn saturating_add(self, duration: Duration) -> Timestamp

Computes self + duration, saturating value on overflow.

See [PrimitiveDateTime::saturating_add] for more implementation details

pub const fn saturating_sub(self, duration: Duration) -> Timestamp

Computes self - duration, saturating value on overflow.

See [PrimitiveDateTime::saturating_sub] for more implementation details