client_sdk::models
Module user

Modules

preferences

Structs

ArchivedExtraUserProfileBits

Archived version of ExtraUserProfileBits, automatically made endian-agnostic. When comparing with the native bitflags, the archived bitflags will truncate any extra bits.

ArchivedFont

ArchivedFriendAddability

ArchivedLocale

ArchivedRelationship

An archived Relationship

ArchivedUser

An archived User

ArchivedUserFlags

Archived version of UserFlags, automatically made endian-agnostic. When comparing with the native bitflags, the archived bitflags will truncate any extra bits.

ArchivedUserPreferences

An archived UserPreferences

ArchivedUserPrefsFlags

Archived version of UserPrefsFlags, automatically made endian-agnostic. When comparing with the native bitflags, the archived bitflags will truncate any extra bits.

ArchivedUserProfile

An archived UserProfile

ArchivedUserProfileBits

Archived version of UserProfileBits, automatically made endian-agnostic. When comparing with the native bitflags, the archived bitflags will truncate any extra bits.

ArchivedUserRelationship

ExtraUserProfileBits

Relationship

RelationshipResolver

The resolver for an archived Relationship

User

UserFlags

NOTE: Remember to clear flag caches when they change

UserPreferences

User preferences

UserPreferencesResolver

The resolver for an archived UserPreferences

UserPrefsFlags

UserProfile

UserProfileBits

UserProfileResolver

The resolver for an archived UserProfile

UserResolver

The resolver for an archived User

Enums

ElevationLevel

Font

FriendAddability

Locale

UserRelationship