client_sdk::models::user
Struct UserFlags

pub struct UserFlags(/* private fields */);

NOTE: Remember to clear flag caches when they change

Implementations

impl UserFlags

pub const BANNED: Self = _

pub const VERIFIED: Self = _

pub const MFA_ENABLED: Self = _

pub const NEEDS_PASSWORD_RESET: Self = _

pub const RESERVED_1: Self = _

pub const RESERVED_2: Self = _

pub const ELEVATION_1: Self = _

pub const ELEVATION_2: Self = _

pub const ELEVATION_3: Self = _

pub const PREMIUM_1: Self = _

pub const PREMIUM_2: Self = _

pub const PREMIUM_3: Self = _

pub const RESERVED_3: Self = _

pub const EXTRA_STORAGE_1: Self = _

pub const EXTRA_STORAGE_2: Self = _

pub const RESERVED_4: Self = _

pub const RESERVED: Self = _

All reserved bits for future use

pub const PRIVATE_FLAGS: Self = _

Always strip these from public responses

pub const ELEVATION: Self = _

elevation level integer

pub const PREMIUM: Self = _

premium level integer

pub const EXTRA_STORAGE: Self = _

extra storage level integer

impl UserFlags

pub const fn empty() -> Self

Get a flags value with all bits unset.

pub const fn all() -> Self

Get a flags value with all known bits set.

pub const fn bits(&self) -> i32

Get the underlying bits value.

The returned value is exactly the bits set in this flags value.

pub const fn from_bits(bits: i32) -> Option<Self>

Convert from a bits value.

This method will return None if any unknown bits are set.

pub const fn from_bits_truncate(bits: i32) -> Self

Convert from a bits value, unsetting any unknown bits.

pub const fn from_bits_retain(bits: i32) -> Self

Convert from a bits value exactly.

pub fn from_name(name: &str) -> Option<Self>

Get a flags value with the bits of a flag with the given name set.

This method will return None if name is empty or doesn’t correspond to any named flag.

pub const fn is_empty(&self) -> bool

Whether all bits in this flags value are unset.

pub const fn is_all(&self) -> bool

Whether all known bits in this flags value are set.

pub const fn intersects(&self, other: Self) -> bool

Whether any set bits in a source flags value are also set in a target flags value.

pub const fn contains(&self, other: Self) -> bool

Whether all set bits in a source flags value are also set in a target flags value.

pub fn insert(&mut self, other: Self)

The bitwise or (|) of the bits in two flags values.

pub fn remove(&mut self, other: Self)

The intersection of a source flags value with the complement of a target flags value (&!).

This method is not equivalent to self & !other when other has unknown bits set. remove won’t truncate other, but the ! operator will.

pub fn toggle(&mut self, other: Self)

The bitwise exclusive-or (^) of the bits in two flags values.

pub fn set(&mut self, other: Self, value: bool)

Call insert when value is true or remove when value is false.

pub const fn intersection(self, other: Self) -> Self

The bitwise and (&) of the bits in two flags values.

pub const fn union(self, other: Self) -> Self

The bitwise or (|) of the bits in two flags values.

pub const fn difference(self, other: Self) -> Self

The intersection of a source flags value with the complement of a target flags value (&!).

This method is not equivalent to self & !other when other has unknown bits set. difference won’t truncate other, but the ! operator will.

pub const fn symmetric_difference(self, other: Self) -> Self

The bitwise exclusive-or (^) of the bits in two flags values.

pub const fn complement(self) -> Self

The bitwise negation (!) of the bits in a flags value, truncating the result.

impl UserFlags

pub const fn iter(&self) -> Iter<UserFlags>

Yield a set of contained flags values.

Each yielded flags value will correspond to a defined named flag. Any unknown bits will be yielded together as a final flags value.

pub const fn iter_names(&self) -> IterNames<UserFlags>

Yield a set of contained named flags values.

This method is like iter, except only yields bits in contained named flags. Any unknown bits, or bits not corresponding to a contained flag will not be yielded.

impl UserFlags

pub const SYSTEM_USER: UserFlags = _

pub const fn from_bits_truncate_public(bits: i32) -> Self

pub const fn elevation(self) -> ElevationLevel

pub const fn with_elevation(self, ev: ElevationLevel) -> Self

pub const fn premium_level(self) -> u8

pub const fn extra_storage_tier(self) -> u8