client_sdk::api
Trait Command

pub trait Command: Sealed {
    type Item: CommandResult;
    type Result: CommandResult;
    type Body: CommandBody;

    const IS_QUERY: bool;
    const HTTP_METHOD: Method;
    const FLAGS: CommandFlags;
    const RATE_LIMIT: RateLimit;
    const SERVER_TIMEOUT: Duration;
    const ROUTE_PATTERN: &'static str;

    // Required methods
    fn format_path<W: Write>(&self, w: W) -> Result;
    fn body(&self) -> &Self::Body;
    fn collect<S, E>(
        stream: S,
    ) -> impl Future<Output = Result<Self::Result, E>> + Send
       where S: Stream<Item = Result<Self::Item, E>> + Send,
             E: From<MissingItemError>;
    fn perms(&self) -> Permissions;

    // Provided methods
    fn path(&self) -> Result<String, Error> { ... }
    fn body_size_hint(&self) -> usize { ... }
    fn add_headers(&self, _map: &mut HeaderMap) { ... }
}

Client Command, tells the client to perform specific requests

A “Command” is a mid-level abstraction around REST endpoints and their bodies. Not perfect, but zero-cost and simple. Other abstractions can be built on top of it.

A command consists of three parts: the URL, the “body”, and headers.

For the case of GET/OPTIONS commands, the body becomes query parameters.

Required Associated Constants

const IS_QUERY: bool

Whether the command has a query string or sends a body

const HTTP_METHOD: Method

HTTP Method used to execute the command

const FLAGS: CommandFlags

Flags for the command, defaults to empty.

const RATE_LIMIT: RateLimit

Baseline rate-limiting parameters, defaults to RateLimit::DEFAULT.

The server may choose to adapt this as needed, and it may not be the only rate-limiting factor depending on the request.

const SERVER_TIMEOUT: Duration

On the server side, how long to wait before timing out the request.

const ROUTE_PATTERN: &'static str

Path pattern for the command (without query) when used with matchit 0.8 or higher.

Required Associated Types

type Item: CommandResult

The underlying type of each returned item, be it one or many.

type Result: CommandResult

Item(s) returned from the server by a given command

type Body: CommandBody

Body type for the command

Required Methods

fn format_path<W: Write>(&self, w: W) -> Result

Serialize/format the REST path (without query)

fn body(&self) -> &Self::Body

Body to be serialized as request body or query parameters (if GET)

fn collect<S, E>( stream: S, ) -> impl Future<Output = Result<Self::Result, E>> + Send
where S: Stream<Item = Result<Self::Item, E>> + Send, E: From<MissingItemError>,

Used to collect the Result from an arbitrary Stream of items.

fn perms(&self) -> Permissions

Computes required permissions

Provided Methods

fn path(&self) -> Result<String, Error>

Serialize/format the REST path (without query) and return it as a string

fn body_size_hint(&self) -> usize

Hint given to preallocate body size, only used for query strings

fn add_headers(&self, _map: &mut HeaderMap)

Insert any additional headers required to perform this command

Object Safety

This trait is not object safe.