client_sdk
Module models

Object data structures used within Lantern

Re-exports

pub use util::fixed_str::FixedStr;

pub use util::thin_str::ThinString;

pub use nullable::Nullable;

pub use aliases::*;

pub use self::asset::*;

pub use self::auth::*;

pub use self::config::*;

pub use self::embed::*;

pub use self::emote::*;

pub use self::file::*;

pub use self::gateway::*;

pub use self::invite::*;

pub use self::message::*;

pub use self::party::*;

pub use self::permission::*;

pub use self::presence::*;

pub use self::role::*;

pub use self::room::*;

pub use self::session::*;

pub use self::sf::*;

pub use self::stats::*;

pub use self::thread::*;

pub use self::user::*;

Modules

aliases

Defines Snowflake aliases to easier keep track of what ID is for what.

asset

auth

config

embed

Message Embed Structures, dual licensed under MIT or Apache 2.0

emote

file

gateway

invite

message

nullable

party

permission

presence

role

room

session

sf

stats

thread

user

util

Structs

Arc

An atomically reference counted shared pointer

FxRandomState2

Lightweight randomly-seeded hash builder for [rustc_hash::FxHasher], uses compile-time random seed that increments on each hash-builder creation.

SmolStr

A SmolStr is a string type that has the following properties:

Snowflake

Snowflake Identifier as used in Lantern Chat.

ThinVec

See the crate’s top level documentation for a description of this type.

Timestamp

UTC Timestamp with nanosecond precision, millisecond-precision when serialized to serde (JSON).

Enums

ArchivedCursor

An archived Cursor

Cursor

Directional search query

CursorResolver

The resolver for an archived Cursor

Type Aliases

EncryptedSnowflake

Encrypted Snowflake, used for obfuscating IDs in URLs.