Module client_sdk::models::embed::v1

Structs

ArchivedBasicEmbedMedia

An archived BasicEmbedMedia

ArchivedEmbedAuthor

An archived EmbedAuthor

ArchivedEmbedField

An archived EmbedField

ArchivedEmbedFooter

An archived EmbedFooter

ArchivedEmbedMedia

An archived EmbedMedia

ArchivedEmbedProvider

An archived EmbedProvider

ArchivedEmbedV1

An archived EmbedV1

BasicEmbedMedia

BasicEmbedMediaResolver

The resolver for an archived BasicEmbedMedia

BoxedEmbedMedia

EmbedAuthor

EmbedAuthorResolver

The resolver for an archived EmbedAuthor

EmbedField

EmbedFieldResolver

The resolver for an archived EmbedField

EmbedFlags

EmbedFooter

EmbedFooterResolver

The resolver for an archived EmbedFooter

EmbedMedia

EmbedMediaResolver

The resolver for an archived EmbedMedia

EmbedProvider

EmbedProviderResolver

The resolver for an archived EmbedProvider

EmbedV1

An embed is metadata taken from a given URL by loading said URL, parsing any meta tags, and fetching extra information from oEmbed sources.

EmbedV1Resolver

The resolver for an archived EmbedV1

NicheEmbedMedia

Rather than implement Archive directly on BoxedEmbedMedia, as all uses are as Option<BoxedEmbedMedia>, use ArchiveWith to achieve niche optimizations

Enums

ArchivedEmbedType

An archived EmbedType

EmbedType

EmbedTypeResolver

The resolver for an archived EmbedType

Traits

VisitMedia

Type Aliases

UrlSignature