client_sdk::models::gateway::message
Enum ClientMsg

#[repr(u8)]
pub enum ClientMsg {
    Heartbeat(HeartbeatPayload),
    Identify(IdentifyPayload),
    Resume(ResumePayload),
    SetPresence(SetPresencePayload),
    Subscribe(SubscribePayload),
    Unsubscribe(UnsubscribePayload),
}

Messages sent from the client to the server

Variants

Heartbeat(HeartbeatPayload)

See new_heartbeat for an easy way to create this message.

Identify(IdentifyPayload)

See new_identify for an easy way to create this message.

Resume(ResumePayload)

See new_resume for an easy way to create this message.

SetPresence(SetPresencePayload)

See new_set_presence for an easy way to create this message.

Subscribe(SubscribePayload)

See new_subscribe for an easy way to create this message.

Unsubscribe(UnsubscribePayload)

See new_unsubscribe for an easy way to create this message.

Implementations

impl ClientMsg

pub const fn opcode(&self) -> ClientMsgOpcode

Returns the discrete opcode for the message

impl ClientMsg

pub fn new_heartbeat() -> Self

Create new Heartbeat message from payload fields.

pub fn new_identify(inner: impl Into<Box<Identify>>) -> Self

Create new Identify message from payload fields.

pub fn new_resume(session: impl Into<Snowflake>) -> Self

Create new Resume message from payload fields.

pub fn new_set_presence(inner: impl Into<Box<SetPresence>>) -> Self

Create new SetPresence message from payload fields.

pub fn new_subscribe(party_id: impl Into<PartyId>) -> Self

Create new Subscribe message from payload fields.

pub fn new_unsubscribe(party_id: impl Into<PartyId>) -> Self

Create new Unsubscribe message from payload fields.