client_sdk::models::gateway::message
Enum ServerMsg

#[repr(u8)]
pub enum ServerMsg {

    Hello(HelloPayload),
    HeartbeatAck(HeartbeatAckPayload),
    Ready(ReadyPayload),
    InvalidSession(InvalidSessionPayload),
    PartyCreate(PartyCreatePayload),
    PartyUpdate(PartyUpdatePayload),
    PartyDelete(PartyDeletePayload),
    RoleCreate(RoleCreatePayload),
    RoleUpdate(RoleUpdatePayload),
    RoleDelete(RoleDeletePayload),
    MemberAdd(MemberAddPayload),
    MemberUpdate(MemberUpdatePayload),
    MemberRemove(MemberRemovePayload),
    MemberBan(MemberBanPayload),
    MemberUnban(MemberUnbanPayload),
    RoomCreate(RoomCreatePayload),
    RoomUpdate(RoomUpdatePayload),
    RoomDelete(RoomDeletePayload),
    RoomPinsUpdate(RoomPinsUpdatePayload),
    MessageCreate(MessageCreatePayload),
    MessageUpdate(MessageUpdatePayload),
    MessageDelete(MessageDeletePayload),
    MessageReactionAdd(MessageReactionAddPayload),
    MessageReactionRemove(MessageReactionRemovePayload),
    MessageReactionRemoveAll(MessageReactionRemoveAllPayload),
    MessageReactionRemoveEmote(MessageReactionRemoveEmotePayload),
    PresenceUpdate(PresenceUpdatePayload),
    TypingStart(TypingStartPayload),
    UserUpdate(UserUpdatePayload),
    ProfileUpdate(ProfileUpdatePayload),
    RelationAdd(RelationAddPayload),
    RelationRemove(RelationRemovePayload),

}

Messages send from the server to the client

Variants

Hello(HelloPayload)

The Hello message initiates the gateway session and expects a ClientMsg::Identify in return.

See new_hello for an easy way to create this message.

HeartbeatAck(HeartbeatAckPayload)

Acknowledgement of a heartbeat

See new_heartbeat_ack for an easy way to create this message.

Ready(ReadyPayload)

See new_ready for an easy way to create this message.

InvalidSession(InvalidSessionPayload)

Sent when the session is no longer valid

See new_invalid_session for an easy way to create this message.

PartyCreate(PartyCreatePayload)

See new_party_create for an easy way to create this message.

PartyUpdate(PartyUpdatePayload)

See new_party_update for an easy way to create this message.

PartyDelete(PartyDeletePayload)

See new_party_delete for an easy way to create this message.

RoleCreate(RoleCreatePayload)

See new_role_create for an easy way to create this message.

RoleUpdate(RoleUpdatePayload)

See new_role_update for an easy way to create this message.

RoleDelete(RoleDeletePayload)

See new_role_delete for an easy way to create this message.

MemberAdd(MemberAddPayload)

See new_member_add for an easy way to create this message.

MemberUpdate(MemberUpdatePayload)

See new_member_update for an easy way to create this message.

MemberRemove(MemberRemovePayload)

See new_member_remove for an easy way to create this message.

MemberBan(MemberBanPayload)

See new_member_ban for an easy way to create this message.

MemberUnban(MemberUnbanPayload)

See new_member_unban for an easy way to create this message.

RoomCreate(RoomCreatePayload)

See new_room_create for an easy way to create this message.

RoomUpdate(RoomUpdatePayload)

See new_room_update for an easy way to create this message.

RoomDelete(RoomDeletePayload)

See new_room_delete for an easy way to create this message.

RoomPinsUpdate(RoomPinsUpdatePayload)

See new_room_pins_update for an easy way to create this message.

MessageCreate(MessageCreatePayload)

See new_message_create for an easy way to create this message.

MessageUpdate(MessageUpdatePayload)

See new_message_update for an easy way to create this message.

MessageDelete(MessageDeletePayload)

See new_message_delete for an easy way to create this message.

MessageReactionAdd(MessageReactionAddPayload)

See new_message_reaction_add for an easy way to create this message.

MessageReactionRemove(MessageReactionRemovePayload)

See new_message_reaction_remove for an easy way to create this message.

MessageReactionRemoveAll(MessageReactionRemoveAllPayload)

See new_message_reaction_remove_all for an easy way to create this message.

MessageReactionRemoveEmote(MessageReactionRemoveEmotePayload)

See new_message_reaction_remove_emote for an easy way to create this message.

PresenceUpdate(PresenceUpdatePayload)

See new_presence_update for an easy way to create this message.

TypingStart(TypingStartPayload)

See new_typing_start for an easy way to create this message.

UserUpdate(UserUpdatePayload)

See new_user_update for an easy way to create this message.

ProfileUpdate(ProfileUpdatePayload)

See new_profile_update for an easy way to create this message.

RelationAdd(RelationAddPayload)

See new_relation_add for an easy way to create this message.

RelationRemove(RelationRemovePayload)

See new_relation_remove for an easy way to create this message.

Implementations

impl ServerMsg

pub const fn opcode(&self) -> ServerMsgOpcode

Returns the discrete opcode for the message

impl ServerMsg

pub fn new_hello(inner: impl Into<Hello>) -> Self

Create new Hello message from payload fields.

The Hello message initiates the gateway session and expects a ClientMsg::Identify in return.

pub fn new_heartbeat_ack() -> Self

Create new HeartbeatAck message from payload fields.

Acknowledgement of a heartbeat

pub fn new_ready(inner: impl Into<Arc<Ready>>) -> Self

Create new Ready message from payload fields.

pub fn new_invalid_session() -> Self

Create new InvalidSession message from payload fields.

Sent when the session is no longer valid

pub fn new_party_create(inner: impl Into<Arc<Party>>) -> Self

Create new PartyCreate message from payload fields.

pub fn new_party_update(inner: impl Into<Arc<PartyUpdateEvent>>) -> Self

Create new PartyUpdate message from payload fields.

pub fn new_party_delete(id: impl Into<PartyId>) -> Self

Create new PartyDelete message from payload fields.

pub fn new_role_create(inner: impl Into<Arc<Role>>) -> Self

Create new RoleCreate message from payload fields.

pub fn new_role_update(inner: impl Into<Arc<Role>>) -> Self

Create new RoleUpdate message from payload fields.

pub fn new_role_delete(inner: impl Into<Arc<RoleDeleteEvent>>) -> Self

Create new RoleDelete message from payload fields.

pub fn new_member_add(inner: impl Into<Arc<PartyMemberEvent>>) -> Self

Create new MemberAdd message from payload fields.

pub fn new_member_update(inner: impl Into<Arc<PartyMemberEvent>>) -> Self

Create new MemberUpdate message from payload fields.

pub fn new_member_remove(inner: impl Into<Arc<PartyMemberEvent>>) -> Self

Create new MemberRemove message from payload fields.

pub fn new_member_ban(inner: impl Into<Arc<PartyMemberEvent>>) -> Self

Create new MemberBan message from payload fields.

pub fn new_member_unban(inner: impl Into<Arc<PartyMemberEvent>>) -> Self

Create new MemberUnban message from payload fields.

pub fn new_room_create(inner: impl Into<Arc<()>>) -> Self

Create new RoomCreate message from payload fields.

pub fn new_room_update(inner: impl Into<Arc<()>>) -> Self

Create new RoomUpdate message from payload fields.

pub fn new_room_delete(inner: impl Into<Arc<RoomDeleteEvent>>) -> Self

Create new RoomDelete message from payload fields.

pub fn new_room_pins_update() -> Self

Create new RoomPinsUpdate message from payload fields.

pub fn new_message_create(inner: impl Into<Arc<RoomMessage>>) -> Self

Create new MessageCreate message from payload fields.

pub fn new_message_update(inner: impl Into<Arc<RoomMessage>>) -> Self

Create new MessageUpdate message from payload fields.

pub fn new_message_delete(inner: impl Into<Arc<MessageDeleteEvent>>) -> Self

Create new MessageDelete message from payload fields.

pub fn new_message_reaction_add( inner: impl Into<Arc<UserReactionEvent>>, ) -> Self

Create new MessageReactionAdd message from payload fields.

pub fn new_message_reaction_remove( inner: impl Into<Arc<UserReactionEvent>>, ) -> Self

Create new MessageReactionRemove message from payload fields.

pub fn new_message_reaction_remove_all() -> Self

Create new MessageReactionRemoveAll message from payload fields.

pub fn new_message_reaction_remove_emote() -> Self

Create new MessageReactionRemoveEmote message from payload fields.

pub fn new_presence_update(inner: impl Into<Arc<UserPresenceEvent>>) -> Self

Create new PresenceUpdate message from payload fields.

pub fn new_typing_start(inner: impl Into<Arc<TypingStart>>) -> Self

Create new TypingStart message from payload fields.

pub fn new_user_update(user: impl Into<Arc<User>>) -> Self

Create new UserUpdate message from payload fields.

pub fn new_profile_update(inner: impl Into<Arc<ProfileUpdateEvent>>) -> Self

Create new ProfileUpdate message from payload fields.

pub fn new_relation_add(inner: impl Into<Arc<Relationship>>) -> Self

Create new RelationAdd message from payload fields.

pub fn new_relation_remove(user_id: impl Into<UserId>) -> Self

Create new RelationRemove message from payload fields.

impl ServerMsg

pub fn matching_intent(&self) -> Option<Intent>

pub fn user_id(&self) -> Option<UserId>

If the event originated from a specific user, get their ID