client_sdk::models
Module party

Structs

ArchivedPartialParty

An archived PartialParty

ArchivedParty

An archived Party

ArchivedPartyFlags

Archived version of PartyFlags, automatically made endian-agnostic. When comparing with the native bitflags, the archived bitflags will truncate any extra bits.

ArchivedPartyMember

An archived PartyMember

ArchivedPartyMemberFlags

Archived version of PartyMemberFlags, automatically made endian-agnostic. When comparing with the native bitflags, the archived bitflags will truncate any extra bits.

ArchivedPartyPreferences

An archived PartyPreferences

ArchivedPartyPrefsFlags

Archived version of PartyPrefsFlags, automatically made endian-agnostic. When comparing with the native bitflags, the archived bitflags will truncate any extra bits.

ArchivedPinFolder

An archived PinFolder

ArchivedPinFolderFlags

Archived version of PinFolderFlags, automatically made endian-agnostic. When comparing with the native bitflags, the archived bitflags will truncate any extra bits.

PartialParty

PartialPartyResolver

The resolver for an archived PartialParty

Party

PartyFlags

PartyMember

PartyMemberFlags

PartyMemberResolver

The resolver for an archived PartyMember

PartyPreferences

PartyPreferencesResolver

The resolver for an archived PartyPreferences

PartyPrefsFlags

PartyResolver

The resolver for an archived Party

PinFolder

PinFolderFlags

PinFolderResolver

The resolver for an archived PinFolder