client_sdk::models
Module presence

Structs

Activity

ActivityResolver

The resolver for an archived Activity

ArchivedActivity

An archived Activity

ArchivedUserPresence

An archived UserPresence

ArchivedUserPresenceFlags

Archived version of UserPresenceFlags, automatically made endian-agnostic. When comparing with the native bitflags, the archived bitflags will truncate any extra bits.

UserPresence

UserPresenceFlags

NOTE: These flags are ordered such that larger values take precedence

UserPresenceResolver

The resolver for an archived UserPresence

Enums

AnyActivity

AnyActivityResolver

The resolver for an archived AnyActivity

ArchivedAnyActivity

An archived AnyActivity